requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
pynput==1.4.2
pyperclip==1.7.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
websocket-client==0.56.0